WEBPAGE/requirement.txt
fastapi
uvicorn
pydantic
python-multipart
google
googlesearch-python
